Implement a basic macro system
This patch introduces two types of macros: symbol macros and function
macros. Symbol macros are used as regular identifiers that always expand
to the same body. Function macros are used as functions and can
interpolate their arguments into the AST that gets created.

There are currently a few limitations, mainly that a macro definition
can't make use of another macro, but that shouldn't be a problem since
macros are not meant to be used for such complicated things.